+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.15.2
+
+- bug fixes
+    - modal: scroll state not restored when the modal container is recreated
+
 ## 1.15.1
 
 - bug fixes